update calendar module

+++ public/html/assets/js/custom.js
@@ -1,10 +1,31 @@
 $(document).ready(function(){
-    // check if google access token is expired or not
-    checkGoogleAccessToken();
 
-    $('.disconnect-yes-btn').click(function() {
+    $('input[name="datetimes"]').daterangepicker({
+        timePicker: true,
+        startDate: moment().startOf('hour'),
+        endDate: moment().startOf('hour').add(32, 'hour'),
+        locale: {
+          format: 'M/DD/YYYY hh:mm A'
+        }
+      });
+
+    $(document).on('click', '.close-calendars-list-modal',function() {
+        $('#calendars-list').modal('hide');
+    })
+
+    var currentURL = window.location.href;
+    console.log('currentURL = ',currentURL);
+    var checkURL = currentURL.includes('/investigator/');
+
+    // console.log(checkURL);
+
+    /* if(checkURL)
+    checkGoogleAccessToken();  */// check if google access token is expired or not
+    
+
+    $('.disconnect-yes-btn').click(function() { // disconnect google sync
         $.ajax({
-            url : 'disconnect-calendar',
+            url : '/investigator/disconnect-calendar',
             method : 'delete',
             success : function(data) {
                 location.reload();
@@ -16,16 +37,19 @@ $(document).ready(function(){
 
 function checkGoogleAccessToken() {
     $.ajax({
-        url : "checkToken",
-        data : {},
+        url : "/investigator/checkToken",
         success : function(data) {
-            console.log(data);
+            // console.log(data);
         }
     });
 }
 
 document.addEventListener('DOMContentLoaded', function() {
 
+    var currentURL = window.location.href;
+    var checkURL = currentURL.includes('/investigator/');
+
+    if(checkURL)
     fetchEvents();
 
     $('.sync-btn').click(function(e) {
@@ -33,13 +57,14 @@ document.addEventListener('DOMContentLoaded', function() {
         var selectedCalendar = $('.select-calendar').val();
 
         $.ajax({
-            url: 'calendar/fetch-events',
+            url: '/investigator/calendar/fetch-events',
             data: {
-                calendarId: selectedCalendar
+                calendar_id: selectedCalendar
             },
             method: 'POST',
             success: function(data) {
                 $('#calendars-list').modal('hide');
+                $('.update-calender-button').removeClass('d-none');
                 var events = JSON.parse(data);
 
                 var source = {
@@ -49,33 +74,51 @@ document.addEventListener('DOMContentLoaded', function() {
                 var calendarEl = document.getElementById('calendar');
                 var calendar = new FullCalendar.Calendar(calendarEl, {
                     initialView: 'dayGridMonth',
-                    events: source
+                    events: source,
+                    dayMaxEvents: 3
                 });
                 calendar.render();
+
+                
             }
         });
 
+    });
+
+
+
+    $('.update-calendar-yes-btn').click(function() {
+        $.ajax({
+            url : 'remove-events',
+            method : 'delete',
+            success : function(data) {
+                location.reload();
+            }
+        })
     });
   
 });
 
 function fetchEvents() {
     $.ajax({
-        url: 'calendar/fetch-events-onload',
+        url: '/investigator/calendar/fetch-events-onload',
         method: 'GET',
         success: function(data) {
             var events = JSON.parse(data);
-console.log(events);
+            if(events.length > 0)
+            $('.update-calender-button').removeClass('d-none');
+
             var source = {
                 events: events
             };
 
             var calendarEl = document.getElementById('calendar');
-            var calendar = new FullCalendar.Calendar(calendarEl, {
-                initialView: 'dayGridMonth',
-                events: source
-            });
-            calendar.render();
+                var calendar = new FullCalendar.Calendar(calendarEl, {
+                    initialView: 'dayGridMonth',
+                    events: source,
+                    dayMaxEvents: 3
+                });
+                calendar.render();
         }
     });
 }
